Implemented the TimerValues component to display a countdown timer for adv-services and chb-services, using localStorage. The time is formatted as minutes:seconds
